4.1.7
- Updated to VS 2019, updated cmakelists to reflect changes required to compile (defaults changed between versions, should still compile same on  vs 2017)
- Ran code analysis, fixing possible overflows and minor code issues.
- Added transform inspector

+++ src/Assets/Image.cpp
@@ -45,7 +45,7 @@ void Image::fill(const glm::uvec4 primaryColor, const glm::uvec4 secondaryColor)
 {
 	if (m_size == glm::ivec2(0))
 		m_size = glm::ivec2(256);
-	const size_t pixelCount = m_size.x * m_size.x;
+	const size_t pixelCount = size_t(m_size.x) * size_t(m_size.x);
 	const size_t componentCount = pixelCount * 4;
 	m_pixelData = new GLubyte[componentCount];
 	m_pitch = m_size.x * 4;
