+++ src/Modules/Graphics/Effects/Skybox.h
@@ -78,7 +78,7 @@ public:
 
 		// Prepare camera index
 		if (m_drawIndex >= m_drawData.size())
-			m_drawData.resize(m_drawIndex + 1);
+			m_drawData.resize(size_t(m_drawIndex) + 1ull);
 		
 		if (m_skyOutOfDate) {
 			convoluteSky(viewport);
