+++ src/Modules/Graphics/Geometry/Prop/Prop_Technique.h
@@ -63,7 +63,7 @@ public:
 		// Exit Early
 		if (m_enabled && m_frameData->viewInfo.size() && m_shapeCube->existsYet() && m_shaderCull->existsYet() && m_shaderGeometry->existsYet()) {
 			if (m_drawIndex >= m_drawData.size())
-				m_drawData.resize(m_drawIndex + 1);
+				m_drawData.resize(size_t(m_drawIndex) + 1ull);
 			auto & drawBuffer = m_drawData[m_drawIndex];
 			auto &camBufferIndex = drawBuffer.bufferCamIndex;
 			auto &propIndexBuffer = drawBuffer.bufferPropIndex;
@@ -144,7 +144,7 @@ public:
 		// Exit Early
 		if (m_enabled && m_frameData->viewInfo.size() && m_shapeCube->existsYet() && m_shaderShadowCull->existsYet() && m_shaderShadowGeometry->existsYet()) {
 			if (m_drawIndex >= m_drawData.size())
-				m_drawData.resize(m_drawIndex + 1);
+				m_drawData.resize(size_t(m_drawIndex) + 1ull);
 			auto & drawBuffer = m_drawData[m_drawIndex];
 			auto &camBufferIndex = drawBuffer.bufferCamIndex;
 			auto &propIndexBuffer = drawBuffer.bufferPropIndex;
