+++ src/Modules/Graphics/Lighting/Point/Point_Technique.h
@@ -61,7 +61,7 @@ public:
 		// Render direct lights	
 		if (m_enabled && m_frameData->viewInfo.size() && m_shapeSphere->existsYet() && m_shader_Lighting->existsYet() && m_shader_Stencil->existsYet()) {
 			if (m_drawIndex >= m_drawData.size())
-				m_drawData.resize(m_drawIndex + 1);
+				m_drawData.resize(size_t(m_drawIndex) + 1ull);
 			auto & drawBuffer = m_drawData[m_drawIndex];
 			auto &camBufferIndex = drawBuffer.bufferCamIndex;
 			auto &lightBufferIndex = drawBuffer.visLights;
